+++ padaria/src/SprinklesDecorator.java
@@ -11,7 +11,7 @@ public class SprinklesDecorator extends Cake {
     }
 
     @Override
-    public double getCost() {
-        return baseCake.getCost() + 2.0;
+    public int getCost() {
+        return baseCake.getCost() + 2;
     }
 }
